LPD-85800 Inline BaseScriptingExecutor logic and remove, no more usages
