+++ modules/apps/portal-scripting/portal-scripting-groovy/src/main/java/com/liferay/portal/scripting/groovy/internal/GroovyScriptingExecutor.java
@@ -5,14 +5,19 @@
 
 package com.liferay.portal.scripting.groovy.internal;
 
+import com.liferay.portal.kernel.scripting.ScriptingContainer;
 import com.liferay.portal.kernel.scripting.ScriptingException;
 import com.liferay.portal.kernel.scripting.ScriptingExecutor;
-import com.liferay.portal.scripting.BaseScriptingExecutor;
+import com.liferay.portal.kernel.util.AggregateClassLoader;
+import com.liferay.portal.kernel.util.FileUtil;
 
 import groovy.lang.Binding;
 import groovy.lang.GroovyRuntimeException;
 import groovy.lang.GroovyShell;
 import groovy.lang.Script;
+
+import java.io.File;
+import java.io.IOException;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -28,9 +33,29 @@ import org.osgi.service.component.annotations.Component;
 	property = "scripting.language=" + GroovyScriptingExecutor.LANGUAGE,
 	service = ScriptingExecutor.class
 )
-public class GroovyScriptingExecutor extends BaseScriptingExecutor {
+public class GroovyScriptingExecutor implements ScriptingExecutor {
 
 	public static final String LANGUAGE = "groovy";
+
+	@Override
+	public void clearCache() {
+	}
+
+	@Override
+	public Map<String, Object> eval(
+			Set<String> allowedClasses, Map<String, Object> inputObjects,
+			Set<String> outputNames, File scriptFile)
+		throws ScriptingException {
+
+		try {
+			String script = FileUtil.read(scriptFile);
+
+			return eval(allowedClasses, inputObjects, outputNames, script);
+		}
+		catch (IOException ioException) {
+			throw new ScriptingException(ioException);
+		}
+	}
 
 	@Override
 	public Map<String, Object> eval(
@@ -82,8 +107,24 @@ public class GroovyScriptingExecutor extends BaseScriptingExecutor {
 	}
 
 	@Override
+	public ScriptingContainer<?> getScriptingContainer() {
+		return null;
+	}
+
+	@Override
 	public ScriptingExecutor newInstance(boolean executeInSeparateThread) {
 		return new GroovyScriptingExecutor();
+	}
+
+	protected ClassLoader getClassLoader() {
+		Class<?> clazz = getClass();
+
+		ClassLoader classLoader = clazz.getClassLoader();
+
+		Thread currentThread = Thread.currentThread();
+
+		return AggregateClassLoader.getAggregateClassLoader(
+			classLoader, currentThread.getContextClassLoader());
 	}
 
 }
